reworked the functions that were being passed as props to PreviewOverlay. Also made some small changes to other props.
