+++ src/components/PreviewOverlay/styles.scss
@@ -1,5 +1,5 @@
 //Mobile portrait
-.container {
+.wrapper {
 	position: fixed; 
 	display: flex;
 	height: 100%;
@@ -16,7 +16,7 @@
 	background-color: green;
 }
 
-	.details {
+	.container {
 		z-index: 2;
 	}
 
@@ -25,7 +25,7 @@
 		}
 
 		.title {
-			font-size: 1.46em;
+			font-size: 1.5em;
 		}
 
 		.controls {
@@ -76,6 +76,7 @@
 		.close {
 			position: fixed;
 			width: 1.62em;
+			height: 1.62em;
 			top: 2.62em;
 			left: 50%;
 			font-size: 1em;
@@ -94,7 +95,7 @@
 
 				.bar {						
 					width: 100%;
-					height: 5px;	
+					height: 5px;	//placeholder
 					position: absolute;
 					top: 50%;
 					left: 50%;
@@ -118,7 +119,7 @@
 
 //Mobile Landscape
 @media (orientation: landscape) {
-	.container {
+	.wrapper {
 		padding-bottom: 1em;
 	}
 
@@ -129,7 +130,7 @@
 
 //Tablet landscape
 @media (min-width: 768px) and (orientation: landscape) {
-	.container {
+	.wrapper {
 		padding-bottom: 3em;
 	}
 
@@ -140,7 +141,7 @@
 
 //Desktop fullscreen
 @media (min-width: 1200px) {
-	.details {
+	.container {
 		top: 80%;
 	}
 }
